moved to network-tests directory, bumped joystream types
