+++ tests/network-tests/package.json
@@ -9,7 +9,7 @@
     "prettier": "prettier --write ./src"
   },
   "dependencies": {
-    "@joystream/types": "^0.6.0",
+    "@joystream/types": "^0.7.0",
     "@polkadot/api": "^0.96.1",
     "@polkadot/keyring": "^1.7.0-beta.5",
     "@types/bn.js": "^4.11.5",
